Fix GoogleNews parser error, reconfigure secrets
This commit fixes a bug introduced in the new "failsafe" if-else
statements in the google news adapter.  We had neglected to pluck
the URL and instead passed entire objects.

This commit also changes how we retreive environment vars/keys
in a way that is hopefully less pattern-specific.  It no longer
relies on `Rails.application.secrets` and uses `ENV['<secret>']`